Add data base

+++ src/main/java/ru/yandex/practicum/shareIt/booking/BookingDto.java
@@ -7,10 +7,10 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class BookingDto {
-    private int id;
+    private Long id;
     private LocalDateTime start;
     private LocalDateTime end;
-    private int item;
-    private int booker;
+    private Long item;
+    private Long booker;
     private BookingStatus status;
 }
