+++ src/main/java/ru/yandex/practicum/shareIt/item/ItemValidator.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public class ItemValidator {
 
-    public static boolean isThereAItem (int itemId, Map<Integer, Item> itemsMap) {
+    public static boolean isThereAItem (long itemId, Map<Long, Item> itemsMap) {
         return itemsMap.containsKey(itemId);
     }
 }
